Changed feature values from floats to integer counts.

+++ src/edu/jhu/thrax/distributional/ContextPhrase.java
@@ -1,6 +1,6 @@
 package edu.jhu.thrax.distributional;
 
-import org.apache.hadoop.io.FloatWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.MapWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.Writable;
@@ -20,13 +20,13 @@ public class ContextPhrase {
 		addFeature(feature_name, 1);
 	}
 
-	public void addFeature(String feature_name, float feature_value) {
+	public void addFeature(String feature_name, int feature_value) {
 		Text feature_text = new Text(feature_name);
 		Writable current_value = features.get(feature_text);
 		if (current_value != null)
-			features.put(feature_text, new FloatWritable(((FloatWritable) current_value).get() + feature_value));
+			features.put(feature_text, new IntWritable(((IntWritable) current_value).get() + feature_value));
 		else
-			features.put(feature_text, new FloatWritable(feature_value));
+			features.put(feature_text, new IntWritable(feature_value));
 	}
 	
 	public Text getPhrase() {
